Towards GUI-36: Organize everything
